Do not use a dedicated uniform domain file
