--- include/ddc/UniformDomain
@@ -1,3 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-#include "ddc/uniform_domain.hpp"
