+++ include/ddc/is_uniform_domain_v
@@ -1,3 +1,3 @@
 // SPDX-License-Identifier: MIT
 
-#include "ddc/uniform_domain.hpp"
+#include "ddc/uniform_discretization.hpp"
